+++ core/XH.ts
@@ -393,28 +393,25 @@ export class XHApi {
     /**
      * Trigger a full reload of the current application.
      *
-     * @param opts - Options for the reload. Can be either:
-     *   - A `string`: Relative path to reload (e.g. 'mobile/').  Defaults to the
-     *       existing location pathname.
-     *   - An `object`: With optional properties:
-     *     - `path` (string): Relative path to reload (e.g. 'mobile/').  Defaults to the
-     *       existing location pathname.
-     *     - `removeQueryParams` (boolean): Whether to remove the existing URL's query parameters.
+     * @param opts - options to govern reload. To support legacy usages, a provided
+     *       string will be treated as `ReloadAppOptions.path`.
      *
      * This method will reload the entire application document in the browser - to trigger a
      * refresh of the loadable content within the app, use {@link refreshAppAsync} instead.
      */
     @action
-    reloadApp(opts?: ReloadAppOptions) {
+    reloadApp(opts?: ReloadAppOptions | string) {
         never().linkTo(this.appLoadModel);
 
+        opts = isString(opts) ? {path: opts} : (opts ?? {});
+
         const {location} = window,
-            path = isString(opts) ? opts : opts?.path,
-            removeQueryParams = isString(opts) ? false : opts?.removeQueryParams,
-            href = path ? `${location.origin}/${path.replace(/^\/+/, '')}` : location.href,
+            href = opts.path
+                ? `${location.origin}/${opts.path.replace(/^\/+/, '')}`
+                : location.href,
             url = new URL(href);
 
-        if (removeQueryParams) url.search = '';
+        if (opts.removeQueryParams) url.search = '';
         // Add a unique query param to force a full reload without using the browser cache.
         url.searchParams.set('xhCacheBuster', Date.now().toString());
         document.location.assign(url);